+++ tests/unitTests/src/unitTestsStorage.js
@@ -8,6 +8,8 @@ import { uploadToGaiaHub, getFullReadUrl,
          getBucketUrl } from '../../../lib/storage/hub'
 import { getFile, encryptContent, decryptContent } from '../../../lib/storage'
 import { BLOCKSTACK_STORAGE_LABEL } from '../../../lib/auth/authConstants'
+import { getPublicKeyFromPrivate } from '../../../lib/keys'
+
 
 class LocalStorage {
   constructor() {
@@ -237,8 +239,9 @@ export function runStorageTests() {
   test('encrypt & decrypt content -- specify key', (t) => {
     t.plan(2)
     const privateKey = '896adae13a1bf88db0b2ec94339b62382ec6f34cd7e2ff8abae7ec271e05f9d8'
+    const publicKey = getPublicKeyFromPrivate(privateKey)
     const content = 'we-all-live-in-a-yellow-submarine'
-    const ciphertext = encryptContent(content, { privateKey })
+    const ciphertext = encryptContent(content, { publicKey })
     t.ok(ciphertext)
     const deciphered = decryptContent(ciphertext, { privateKey })
     t.equal(content, deciphered)
@@ -306,6 +309,57 @@ export function runStorageTests() {
 
     FetchMock.get(fullReadUrl, encryptContent(fileContent))
     const encryptOptions = { encrypt: true }
+    const decryptOptions = { decrypt: true }
+    // put and encrypt the file
+    putFile(path, fileContent, encryptOptions)
+      .then((publicURL) => {
+        t.ok(publicURL, fullReadUrl)
+      }).then(() => {
+        // read and decrypt the file
+        getFile(path, decryptOptions).then(readContent => {
+          t.equal(readContent, fileContent)
+          // put back whatever was inside before
+          window.localStorage.setItem(BLOCKSTACK_STORAGE_LABEL, oldContent)
+        })
+      })
+  })
+
+  test('putFile encrypt using specifying public key & getFile decrypt', (t) => {
+    t.plan(2)
+
+    const privateKey = 'a5c61c6ca7b3e7e55edee68566aeab22e4da26baa285c7bd10e8d2218aa3b229'
+    const publicKey = getPublicKeyFromPrivate(privateKey)
+
+    // save any previous content
+    const oldContent = window.localStorage.getItem(BLOCKSTACK_STORAGE_LABEL)
+    const userData = JSON.stringify({ appPrivateKey: privateKey })
+    // simulate fake user signed in
+    window.localStorage.setItem(BLOCKSTACK_STORAGE_LABEL, userData)
+
+    const path = 'file.json'
+    const gaiaHubConfig = {
+      address: '1NZNxhoxobqwsNvTb16pdeiqvFvce3Yg8U',
+      server: 'https://hub.blockstack.org',
+      token: '',
+      url_prefix: 'gaia.testblockstack.org/hub/'
+    }
+
+    const fullReadUrl = 'https://gaia.testblockstack.org/hub/1NZNxhoxobqwsNvTb16pdeiqvFvce3Yg8A/file.json'
+    const fileContent = JSON.stringify({ test: 'test' })
+
+    const getOrSetLocalGaiaHubConnection = sinon.stub().resolves(gaiaHubConfig)
+    const uploadToGaiaHub = sinon.stub().resolves(fullReadUrl) // eslint-disable-line no-shadow
+    const getFullReadUrl = sinon.stub().resolves(fullReadUrl) // eslint-disable-line no-shadow
+
+    const { putFile } = proxyquire('../../../lib/storage', {
+      './hub': { getOrSetLocalGaiaHubConnection, uploadToGaiaHub }
+    })
+    const { getFile } = proxyquire('../../../lib/storage', { // eslint-disable-line no-shadow
+      './hub': { getOrSetLocalGaiaHubConnection, getFullReadUrl }
+    })
+
+    FetchMock.get(fullReadUrl, encryptContent(fileContent))
+    const encryptOptions = { encrypt: publicKey }
     const decryptOptions = { decrypt: true }
     // put and encrypt the file
     putFile(path, fileContent, encryptOptions)